Check if user exists

+++ libs/helpers.js
@@ -57,22 +57,7 @@ helpers.controllerMethods = function( req, res, acceptableMethods = ['GET','POST
     }    
 }
 
-// Generate unique id
-helpers.uuid = function() {
-    // Token size
-    const COUNT = 20;
-    // String path
-    var _sym = 'abcdefghijklmnopqrstuvwxyz1234567890';
-    // Return
-    var str  = '';
-    // Generate string
-    for(var i = 0; i < COUNT; i++)
-        str += _sym[parseInt(Math.random() * (_sym.length))];
-
-    return str;
-}
-
-// create file
+// File creation
 helpers.fileAdd = function( file, data, dir, callback ){
     if( !file && !data && !dir ) {
         // Error
@@ -92,18 +77,73 @@ helpers.fileAdd = function( file, data, dir, callback ){
                 callback("Error creating path " + fullDir );
             }
         }
-        // Create file
-        fs.writeFile( fullDir +"/"+ file + ".json", JSON.stringify(data), function(err) {
-            if(err) {
-                // Error writing file
-                callback("Error creating file " + err );
-            } else {
-                // File writed
-                callback(false);
-            }
-        });         
+
+        // Check if file exists
+        if( !helpers.fileExists( file, "users") ){
+            // Create file
+            fs.writeFile( fullDir +"/"+ file + ".json", JSON.stringify(data), function(err) {
+                if(err) {
+                    // Error writing file
+                    callback("Error creating file " + err );
+                } else {
+                    // File written
+                    callback(false);
+                }
+            });         
+        } else {
+            // File exist
+            callback("Error creating file. User file exists");
+        }
     }
 }
+
+// File check if exists
+helpers.fileExists = function( file, dir ){
+    
+    // Path
+    const fullDir = path.join(__dirname,"../.data/" + dir);
+    // File
+    const fullFile =  fullDir +"/"+ file + ".json";
+    // Check if exists
+    return ( fs.existsSync(fullFile) ? true : false );
+}
+
+// List files and check if email exists
+helpers.userExists = function( userEmail ){
+    // Control
+    var userExist = false;
+    // Path
+    const fullDir = path.join(__dirname,"../.data/users");
+    // List files
+    fs.readdirSync(fullDir).forEach(file => {
+        // Read user file
+        try{
+            var { email } = JSON.parse(fs.readFileSync(fullDir +"/"+ file));
+            if( userEmail == email ) userExist = true;
+        } catch( err ){}
+    });
+
+    // Users doesn't exist
+    return userExist;
+}
+
+// Generate unique id
+helpers.uuid = function() {
+    // Token size
+    const COUNT = 20;
+    // String path
+    var _sym = 'abcdefghijklmnopqrstuvwxyz1234567890';
+    // Return
+    var id  = '';
+    // Generate string
+    for(var i = 0; i < COUNT; i++)
+    id += _sym[parseInt(Math.random() * (_sym.length))];
+
+    // Check if uuid exists // Recursive check if ID was used
+    if( helpers.fileExists(id,"users") )        helpers.uuid();
+    else                                        return id;
+}
+
 
 
 // Export
